+++ providers/standard/provider.yaml
@@ -119,7 +119,6 @@ hooks:
 triggers:
   - integration-name: Standard
     python-modules:
-      - airflow.providers.standard.triggers.asset
       - airflow.providers.standard.triggers.external_task
       - airflow.providers.standard.triggers.file
       - airflow.providers.standard.triggers.temporal
