+++ providers/standard/src/airflow/providers/standard/exceptions.py
@@ -57,14 +57,6 @@ class DuplicateStateError(AirflowExternalTaskSensorException):
     """Raised when duplicate states are provided across allowed, skipped and failed states."""
 
 
-class AssetEventSensorException(AirflowException):
-    """Base exception for all AssetEventSensor related errors."""
-
-
-class UnexpectedAssetEventTriggerEventError(AssetEventSensorException):
-    """Raised when the AssetEventTrigger returns an unexpected event to the sensor."""
-
-
 class HITLTriggerEventError(Exception):
     """Raised when TriggerEvent contains error."""
